requirements.txt
Flask==2.2.2
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12